Move utils to there own file
